Move assertion handler to daemon/common, wire into cc_binary
- Moved assertion_handler.{hpp,cpp} from src/testing/ to
  src/daemon/src/common/ so the same handler can be used by both
  unit tests and the daemon binary
- Replaced #pragma once with #ifndef header guard per codebase convention
- Removed testonly = True from the new target so the daemon can link it
- Added the new target as a dep to the daemon cc_binary; removed the
  duplicate inline handler from main.cpp
- Updated lm_cc_test macro to reference the new Bazel label
