--- score/launch_manager/src/testing/BUILD
@@ -1,25 +1,0 @@
-# *******************************************************************************
-# Copyright (c) 2026 Contributors to the Eclipse Foundation
-#
-# See the NOTICE file(s) distributed with this work for additional
-# information regarding copyright ownership.
-#
-# This program and the accompanying materials are made available under the
-# terms of the Apache License Version 2.0 which is available at
-# https://www.apache.org/licenses/LICENSE-2.0
-#
-# SPDX-License-Identifier: Apache-2.0
-# *******************************************************************************
-load("@rules_cc//cc:defs.bzl", "cc_library")
-
-cc_library(
-    name = "testing_assertion_handler",
-    testonly = True,
-    srcs = ["assertion_handler.cpp"],
-    hdrs = ["assertion_handler.hpp"],
-    include_prefix = "score/mw/launch_manager/testing",
-    strip_include_prefix = "/score/launch_manager/src/testing",
-    visibility = ["//score/launch_manager:__subpackages__"],
-    deps = ["@score_baselibs//score/language/futurecpp"],
-    alwayslink = True,
-)
